LayoutTests/fullscreen/full-screen-keyboard-lock-option.html

<!DOCTYPE html>
<html>
<head>
    <title>This tests that full screen mode with keyboard lock option allows ESC key event.</title>
</head>
<body>
    <div width: 100%; height: 100%;" id="target">
        <p>Test passes if ESC key received by the test case after enter to fullscreen.</p>
    </div>
    <script src="full-screen-test.js"></script>
    <script>
    window.internals.settings.setFullScreenKeyboardLock(false);

    runWithKeyDown(async () => {
        try {
            const target = document.querySelector("#target");
            await target.requestFullscreen({ keyboardLock: "browser" });

            // Wait for fullscreen change event
            await new Promise((resolve) => {
                waitForEvent(document, "fullscreenchange", resolve, /*endit=*/ false, /*once=*/ true);
            });

            testExpected("document.fullscreenElement !== null", true);

            // Wait for Escape key press and test inside the callback
            await new Promise((resolve) => {
                waitForEvent(document, "keypress", (event) => {
                    testExpected("event.keyCode", 27); // 'Escape'
                    resolve();
                }, /*endit=*/ false, /*once=*/ true);
                eventSender.keyDown("escape");
            });

        } catch (err) {
            logResult(false, `An error occurred while trying to enter full-screen mode: ${err.message}`);
        } finally {
            endTest();
        }
    });
    </script>
</body>
</html>
