LayoutTests/fullscreen/full-screen-keyboard-lock-option-expected.txt

FAIL: Timed out waiting for notifyDone to be called

Test passes if ESC key received by the test case after enter to fullscreen.

EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement !== null == 'true') OK
